Novos campos e docs para 0.3.0
Este commit altera a versão de 0.2.0 para 0.3.0.

Alterações:

    - Move a página web de jultty.github.io para este repositório
    - Dá prioridade à página Netlify em todos os links
    - Atualiza versão da dependência spring-boot-starter-test para 2.6.7
    - Passa a carregar dados iniciais pelo arquivo data.sql
    - Adiciona campos para tag e para links das páginas e tags no Netlify
    - Cria e atualiza getters e setters que constróem as urls
    - Atualiza constructors, equals, toString e hashCode
    - Muda tipo da variável Id para Integer
    - Comenta as configurações do arquivo application.properties
    - Adiciona configurações do JPA ao arquivo application.properties
    - O back-end agora serve páginas HTML básicas para substituir erros whitelabel

+++ src/main/java/com/semtec/api/TermoController.java
@@ -48,7 +48,7 @@ class TermoController {
     }
 
     @GetMapping("/termo/{id}")
-    EntityModel<Termo> exibeTermo(@PathVariable Long id) {
+    EntityModel<Termo> exibeTermo(@PathVariable Integer id) {
 
         Termo termo = repository.findById(id)
                 .orElseThrow(() -> new TermoNotFoundException(id));
@@ -58,7 +58,7 @@ class TermoController {
 
     @PutMapping("/termo/{id}")
     Termo substituiTermo(@RequestBody Termo
-                         novoTermo, @PathVariable Long id) {
+                         novoTermo, @PathVariable Integer id) {
 
         return repository.findById(id)
                 .map(termo -> {
@@ -73,7 +73,7 @@ class TermoController {
     }
 
     @DeleteMapping("/termo/{id}")
-    void apagaTermo(@PathVariable Long id) {
+    void apagaTermo(@PathVariable Integer id) {
         repository.deleteById(id);
     }
 }
