+++ src/main/java/com/semtec/api/TermoNotFoundException.java
@@ -1,7 +1,7 @@
 package com.semtec.api;
 
 public class TermoNotFoundException extends RuntimeException {
-    TermoNotFoundException(Long id) {
+    TermoNotFoundException(Integer id) {
         super("Termo não encontrado com o ID " + id);
     }
 }
