+++ src/main/java/com/semtec/api/TermoRepository.java
@@ -2,6 +2,6 @@ package com.semtec.api;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface TermoRepository extends JpaRepository<Termo, Long> {
+interface TermoRepository extends JpaRepository<Termo, Integer> {
 
 }
